Optimizations, corrections, cleanup, debug enhancements and feature extensions
Feature extensions:
* method to get raw version of the port init value
* Add a signal that notifies when complete write occurs in one sweep (such as when apx server connect is completed)

Debug enhancements:
* Add debug messages if write fails
* add apx file testcase
* and improve socket state debugging
* Update the unittest for remotefile so that it is possible to execute

Corrections:
* fix issue for large apxProvidePortConnect and add a todo for another potential issue in apxDataMsg
* Remove the FileManagerWorker since only limited work was offloaded to it and it resulted in issues with multiple threads accessing the same socket. Should consider offloading all encoding+decoding to separate thread later.
* add return value check of socket write
* fix issue where decode32 with longbit but insufficient data was interpreted as a parse error
* manage data that may arrive while we are processing previous data (not sure if Qt would signal if data is added wile processin readyread() - anyway this reduces latency)

Optimizations:
* Optimizations removing magic numbers and reducing use of temporary pointers
* optimize data/memory allocation
* make it possible to desire in data notifications using id only
* Send signals with const ref (Warning: breaks backwards compatablity for previous signal+slot connections)
* optimize datastructure lookup
* const and other optimization updates aiding the compiler
* optimize processing of require signals

Cleanup:
* qstring cleanup
* Refactor assert to not cause gcc compiler warning in production build where Q_ASSERTS are removed
* use unique define for APX unit tests to avoid issue with rmf mock socket
* add Q_ASSERT of issue that can not occur in runtime
* remove unused RMF_CMD_* redefinitions and includes
* include cleanup
* remove unused define
* correct include guard and compiler warning

+++ remotefile/test/test_qrmf_proto.cpp
@@ -134,7 +134,7 @@ void TestRemoteFileProtocol::test_packFileInfo()
    int result;
 
    result = RemoteFile::packFileInfo(array.data(), array.length(), file1);
-   QCOMPARE(result, RMF_FILEINFO_BASE_LEN+(int)strlen(name1)+1);
+   QCOMPARE(result, (int)RMF_FILEINFO_BASE_LEN+(int)strlen(name1)+1);
 
    QByteArray expected = QByteArray("\x03\x00\x00\x00\x78\x56\x34\x12\xe8\x03\x00\x00\x00\x00\x00\x00" //16
                                     "\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00" //32
@@ -155,7 +155,7 @@ void TestRemoteFileProtocol::test_unpackFileInfo()
                                     ,RMF_FILEINFO_BASE_LEN+(int)strlen(name1)+1);
    RemoteFile::File file1;
    int result = RemoteFile::unpackFileInfo(array.constData(), array.constData()+array.length(), file1);
-   QCOMPARE(result, RMF_FILEINFO_BASE_LEN+(int)strlen(name1)+1);
+   QCOMPARE(result, (int)RMF_FILEINFO_BASE_LEN+(int)strlen(name1)+1);
    QCOMPARE(file1.mName, QString("file1.txt"));
    QCOMPARE(file1.mAddress, (quint32) 0x12345678);
    QCOMPARE(file1.mLength, (quint32) 1000);
@@ -164,7 +164,7 @@ void TestRemoteFileProtocol::test_unpackFileInfo()
 
    //it should also work if null-terminator is forgotten
    result = RemoteFile::unpackFileInfo(array.constData(), array.constData()+array.length()-1, file1);
-   QCOMPARE(result, RMF_FILEINFO_BASE_LEN+(int)strlen(name1));
+   QCOMPARE(result, (int)RMF_FILEINFO_BASE_LEN+(int)strlen(name1));
    QCOMPARE(file1.mName, QString("file1.txt"));
    QCOMPARE(file1.mAddress, (quint32) 0x12345678);
    QCOMPARE(file1.mLength, (quint32) 1000);
@@ -173,7 +173,7 @@ void TestRemoteFileProtocol::test_unpackFileInfo()
 
    //what if array cuts of directly after RMF_FILEINFO_BASE_LEN bytes?
    result = RemoteFile::unpackFileInfo(array.constData(), array.constData()+RMF_FILEINFO_BASE_LEN, file1);
-   QCOMPARE(result, RMF_FILEINFO_BASE_LEN);
+   QCOMPARE(result, (int)RMF_FILEINFO_BASE_LEN);
    QCOMPARE(file1.mName, QString(""));
    QCOMPARE(file1.mAddress, (quint32) 0x12345678);
    QCOMPARE(file1.mLength, (quint32) 1000);
@@ -186,7 +186,7 @@ void TestRemoteFileProtocol::test_unpackFileInfo()
                          "\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00" //48
                          "\x00",RMF_FILEINFO_BASE_LEN+1); //49
    result = RemoteFile::unpackFileInfo(array.constData(), array.constData()+array.length(), file1);
-   QCOMPARE(result, RMF_FILEINFO_BASE_LEN+1);
+   QCOMPARE(result, (int)RMF_FILEINFO_BASE_LEN+1);
    QCOMPARE(file1.mName, QString(""));
    QCOMPARE(file1.mAddress, (quint32) 0x12345678);
    QCOMPARE(file1.mLength, (quint32) 1000);
@@ -230,12 +230,12 @@ void TestRemoteFileProtocol::test_unpackFileOpen()
       QCOMPARE(result, 0);
    }
    result = RemoteFile::unpackFileOpen(array.constData(),array.constData()+array.length(),address);
-   QCOMPARE(result, RMF_FILE_OPEN_LEN);
+   QCOMPARE(result, (int)RMF_FILE_OPEN_LEN);
    QCOMPARE(address, (quint32)0x00000000);
 
    array = QByteArray(  "\x0a\x00\x00\x00\x78\x56\x34\x12",RMF_FILE_OPEN_LEN);
    result = RemoteFile::unpackFileOpen(array.constData(),array.constData()+array.length(),address);
-   QCOMPARE(result, RMF_FILE_OPEN_LEN);
+   QCOMPARE(result, (int)RMF_FILE_OPEN_LEN);
    QCOMPARE(address, (quint32)0x12345678);
 
    //verify invalid cmdType check
@@ -279,12 +279,12 @@ void TestRemoteFileProtocol::test_unpackFileClose()
       QCOMPARE(result, 0);
    }
    result = RemoteFile::unpackFileClose(array.constData(),array.constData()+array.length(),address);
-   QCOMPARE(result, RMF_FILE_CLOSE_LEN);
+   QCOMPARE(result, (int)RMF_FILE_CLOSE_LEN);
    QCOMPARE(address, (quint32)0x00000000);
 
    array = QByteArray(  "\x0b\x00\x00\x00\x78\x56\x34\x12",RMF_FILE_CLOSE_LEN);
    result = RemoteFile::unpackFileClose(array.constData(),array.constData()+array.length(),address);
-   QCOMPARE(result, RMF_FILE_CLOSE_LEN);
+   QCOMPARE(result, (int)RMF_FILE_CLOSE_LEN);
    QCOMPARE(address, (quint32)0x12345678);
 
    //verify invalid cmdType check
